+++ hustack-backend/baseweb/src/main/java/com/hust/baseweb/applications/exam/model/response/ExamStudentResultDetailsRes.java
@@ -9,7 +9,7 @@ public interface ExamStudentResultDetailsRes {
     String getEmail();
     String getPhone();
     String getExamResultId();
-    Integer getTotalScore();
+    Float getTotalScore();
     Integer getTotalTime();
     String getSubmitedAt();
 }
